+++ src/test/test_compiler.sh
@@ -65,6 +65,7 @@ TEST_FILES=(
     "src/test-resources/test_narrowing_warn.c"
     "src/test-resources/test_func_identifier.c"
     "src/test-resources/test_many_params_locals.c"
+    "src/test-resources/test_ptr_precedence.c"
 )
 
 mkdir -p build/test